requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
